AVISPA_May21/requirements.txt
absl-py==0.11.0
astor==0.8.1
blas==1.0
c-ares==1.16.1
ca-certificates==2020.10.14
certifi==2020.6.20
cloudpickle==1.6.0
cudatoolkit==10.1.243
cudnn==7.6.5
cupti==10.1.168
cycler==0.10.0
cytoolz==0.11.0
dask-core==2.30.0
decorator==4.4.2
freetype==2.10.4
gast==0.4.0
google-pasta==0.2.0
grpcio==1.14.1
h5py==2.8.0
hdf5==1.10.2
imageio==2.9.0
importlib-metadata==2.0.0
intel-openmp==2019.4
jpeg==9b
keras-applications==1.0.8
keras-base==2.3.1
keras-gpu==2.3.1
keras-preprocessing==1.1.0
kiwisolver==1.3.0
lcms2==2.11
libgcc-ng==9.1.0
libgfortran-ng==7.3.0
libpng==1.6.37
libprotobuf==3.13.0.1
libstdcxx-ng==9.1.0
libtiff==4.1.0
markdown==3.3.3
matplotlib-base==3.3.2
matplotlib==3.3.2
mkl==2018.0.3
mkl_fft==1.0.6
mkl_random==1.0.1
networkx==2.5
numpy-base==1.15.4
numpy==1.15.4
olefile==0.46
openssl==1.0.2u
pandas==0.25.0
pillow==8.0.1
pip==20.2.4
protobuf==3.13.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywavelets==1.1.1
pyyaml==5.3.1
scikit-image==0.16.2
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.11.0
setuptools==50.3.1
six==1.15.0
tbb==2020.3
tensorboard==1.14.0
tensorflow-base==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow==1.14.0
termcolor==1.1.0
toolz==0.11.1
tornado==6.0.4
werkzeug==1.0.1
wheel==0.35.1
wrapt==1.12.1
xz==5.2.5
zipp==3.4.0
zlib==1.2.11
zstd==1.4.5